requirements.txt
asgiref==3.3.1
Django==3.1.4
python-slugify==4.0.1
pytz==2020.4
sqlparse==0.4.1
text-unidecode==1.3
